disabled state and clean files
